improved deserialisation schema
